requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
Pillow==12.0.0
python-decouple==3.8
django-extensions==4.1
django-filter==25.2
django-import-export==4.3.12
djongo==1.3.6
pymongo==4.6.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0